requirements.txt
BeautifulSoup4
streamlit
langchain
python-dotenv
faiss-cpu
google-generativeai
streamlit-chat